\name{cosso-internal}
\alias{rescale}
\alias{genK}
\alias{bigGram}
\alias{wsGram}
\alias{sspline}
\alias{cvsplit}
\alias{gcvlam}
\alias{twostep}
\alias{cvadd}
\alias{bicadd}
\alias{gcv}
\alias{bic}
\alias{compdf}
\alias{rho}
\alias{df.est}
\alias{cvM.qr}
\alias{cvLam.qr}
\alias{CVadd}
\alias{cquan}
\alias{garrote.qr}
\alias{kqr}
\alias{cvsplitID}
\alias{My_solve}
\alias{My_solve.QP}
\alias{ACV.lambda}
\alias{twostep.Cox}
\alias{sspline.Cox}
\alias{gradient.Hessian.C}
\alias{garrote.Cox}
\alias{gradient.Hessian.Theta}
\alias{PartialLik}
\alias{RiskSet}

\title{Internal Functions}

\description{
  Internal cosso functions.
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }