\name{cosso-internal}
\alias{rescale}
\alias{genK}
\alias{genK.cat}
\alias{bigGram}
\alias{wsGram}
\alias{cvsplitID}
\alias{My_solve}
\alias{My_solve.QP}

\alias{cosso.Gaussian}
\alias{tune.cosso.Gaussian}
\alias{twostep.Gaussian}
\alias{sspline}
\alias{cvlam.Gaussian}
\alias{SSANOVAwt.Gaussian}

\alias{cosso.Binomial}
\alias{tune.cosso.Binomial}
\alias{cvlam.logistic}
\alias{twostep.Binomial}
\alias{garrote.Logistic}
\alias{garrote.Logistic.GH}
\alias{sspline.Binomial}
\alias{SSANOVAwt.Binomial}

\alias{cosso.qr}
\alias{cosso.qr.Parallel}
\alias{cosso.qr.Sequential}
\alias{twostep.qr}
\alias{kqr}
\alias{garrote.qr}
\alias{rho}
\alias{SSANOVAwt.qr}

\alias{cosso.Cox}
\alias{cosso.Cox.Parallel}
\alias{cosso.Cox.Sequential}
\alias{tune.cosso.Cox}
\alias{ACV.lambda}
\alias{twostep.Cox}
\alias{sspline.Cox}
\alias{gradient.Hessian.C}
\alias{garrote.Cox}
\alias{gradient.Hessian.Theta}
\alias{SSANOVAwt.Cox}
\alias{PartialLik}
\alias{RiskSet}

\title{Internal Functions}

\description{
  Internal cosso functions.
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }